add volume minted

+++ internal/entity/users.go
@@ -31,10 +31,10 @@ type FilteArtist struct {
 }
 
 type UserStats struct {
-	CollectionCreated int32 `bson:"collection_created" json:"collection_created"`
-	NftMinted         int32 `bson:"nft_minted" json:"nft_minted"`
-	OutputMinted      int32 `bson:"output_minted" json:"output_minted"`
-	VolumeMinted      int32 `bson:"volume_minted" json:"volume_minted"`
+	CollectionCreated int32   `bson:"collection_created" json:"collection_created"`
+	NftMinted         int32   `bson:"nft_minted" json:"nft_minted"`
+	OutputMinted      int32   `bson:"output_minted" json:"output_minted"`
+	VolumeMinted      float64 `bson:"volume_minted" json:"volume_minted"`
 }
 
 type Users struct {
